+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v0.5
 - Support min/max for number type, fix #91
 - Support attributes and input className customization, fix #76
 - Pass events to inputs, fix #53, #67
+- Set class on form-group div indicating its depth within form, fix #64
 
 v0.4.8